LayoutTests/platform/chromium-win/websocket/tests/bad-sub-protocol-expected.txt

CONSOLE MESSAGE: Wrong protocol for WebSocket ''
CONSOLE MESSAGE: Wrong protocol for WebSocket '\u0000'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\u0009'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\u001B'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\u007F'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\u0080'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\u3042'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\uFFFF'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\uFEFF'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\uFFFE'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\uD840\uDC0B'
Test WebSocket bad sub-protocol names.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "	") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "
") threw exception SyntaxError: Unexpected token ILLEGAL.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "
") threw exception SyntaxError: Unexpected token ILLEGAL.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "あ") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "￿") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "﻿") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "￾") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "𠀋") threw exception Error: SYNTAX_ERR: DOM Exception 12.
